backend/requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
sqlalchemy[asyncio]~=2.0.37
asyncpg
python-dotenv~=1.0.1
python-jose[cryptography]
passlib[bcrypt]~=1.7.4

pydantic~=2.10.6